chore: 상수, 타입 파일 분리
